==> 03A_CVSummaryPlot.Rmd <==
---
title: "combining tpm and prediction for all samples"
output: html_document
date: "2025-02-18"
---

# Set up
```{r}
rm(list=ls())
library(tidyverse)
source("~/Documents/Projects/AnalyticalValidation/AV_Plots.R")
```

# Read in counts
```{r}
AllCounts<-read.csv("/Users/alinaklineschoder/Desktop/Data_AnalyticalValidation_Organized/FinalCounts.csv")%>%
  dplyr::select(-c("X"))%>%
  column_to_rownames(var="gene")

SampleInfo<-read.csv("/Users/alinaklineschoder/Desktop/AVSampleKey.csv")%>%
  filter(Sample.Name!="CP3_Batch_N16")

AllCounts<-AllCounts[include,]
```

### Table 1. FT 1 vs FT2
```{r}
FT_Meta<-SampleInfo%>%filter(Category=="RNAStabilityFT")%>%
  dplyr::select(Sample.Name,Control,Parameter)
FT_Counts<-AllCounts[,FT_Meta$Sample.Name]%>%
  rownames_to_column(var="Gene")%>%
  pivot_longer(cols=setdiff(colnames(.),"Gene"),names_to = "Sample.Name")

FT<-left_join(FT_Counts,FT_Meta)%>%
  dplyr::select(-c(Sample.Name))%>%
  pivot_wider(names_from = Parameter,values_from = value)%>%
  rowwise()%>%
  mutate(FT=100*sd(c(`2FT`,`1FT`))/mean(c(`2FT`,`1FT`)))%>%
  select(Gene,Control,FT)%>%
  mutate(Type="RNA Stability - Freeze/Thaw")

Tab<-FT%>%
  pivot_wider(names_from = Control,values_from = FT)
```

### Table 2. Over 6mo/12mo
```{r}
Time_Meta<-SampleInfo%>%filter(Category=="RNAStabilityTime")%>%
  dplyr::select(Sample.Name,Control,Parameter)

Time_Counts<-AllCounts[,Time_Meta$Sample.Name]%>%
  rownames_to_column(var="Gene")%>%
  pivot_longer(cols=setdiff(colnames(.),"Gene"),names_to = "Sample.Name")

Time<-left_join(Time_Counts,Time_Meta)%>%
  dplyr::select(-c(Sample.Name))%>%
  pivot_wider(names_from = Parameter,values_from = value)%>%
  rowwise()%>%
  mutate(`6mo`=100*sd(c(T0,`5D`,`17D`,`1M`,`3M`,`6M`))/mean(c(T0,`5D`,`17D`,`1M`,`3M`,`6M`)),
         `12mo`=100*sd(c(T0,`5D`,`17D`,`1M`,`3M`,`6M`,`12M`))/mean(c(T0,`5D`,`17D`,`1M`,`3M`,`6M`,`12M`)))%>%
  select(Gene,`6mo`,`12mo`,Control)%>%
  mutate(Type="RNA Stability - Time")

Tab<-Time%>%
  select(-c(`6mo`))%>%
  pivot_wider(names_from = Control,values_from = `12mo`)
```

### Table 3. +/- DNASE
```{r}
DNAse_Meta<-SampleInfo%>%filter(Category=="Interference")%>%
  dplyr::select(Sample.Name,Control,Parameter)

DNAse_Counts<-AllCounts[,DNAse_Meta$Sample.Name]%>%
  rownames_to_column(var="Gene")%>%
  pivot_longer(cols=setdiff(colnames(.),"Gene"),names_to = "Sample.Name")

DNAse<-left_join(DNAse_Counts,DNAse_Meta)%>%
  dplyr::select(-c(Sample.Name))%>%
  pivot_wider(names_from = Parameter,values_from = value)%>%
  rowwise()%>%
  mutate(DNASE=100*sd(c(W,WO))/mean(c(W,WO)))%>%
  select(Gene,DNASE,Control)%>%
  mutate(Type="Interference")

Tab<-DNAse%>%
  pivot_wider(names_from = Control,values_from = DNASE)
```

### Table 4. RNA Input
```{r}
RI_Meta<-SampleInfo%>%filter(Category=="RNAInput")%>%
  dplyr::select(Sample.Name,Control,Parameter)

RI_Counts<-AllCounts[,RI_Meta$Sample.Name]%>%
  rownames_to_column(var="Gene")%>%
  pivot_longer(cols=setdiff(colnames(.),"Gene"),names_to = "Sample.Name")

RI<-left_join(RI_Counts,RI_Meta)%>%
  dplyr::select(-c(Sample.Name))%>%
  pivot_wider(names_from = Parameter,values_from = value)%>%
  rowwise()%>%
  mutate(RI_400_500=100*sd(c(RI400,RI500))/mean(c(RI400,RI500)),
         RI_500_600=100*sd(c(RI600,RI500))/mean(c(RI600,RI500)),
         All=100*sd(c(RI400,RI600,RI500))/mean(c(RI400,RI600,RI500)))%>%
  select(Gene,RI_400_500,RI_500_600,All,Control)%>%
  mutate(Type="RNA Input")

Tab<-RI%>%
  select(-c(All,RI_400_500))%>%
  pivot_wider(names_from = Control,values_from = RI_500_600)
```
### Table 5. Library Input
```{r}
LI_Meta<-SampleInfo%>%filter(Category=="LibraryInput")%>%
  dplyr::select(Sample.Name,Control,Parameter)

LI_Counts<-AllCounts[,LI_Meta$Sample.Name]%>%
  rownames_to_column(var="Gene")%>%
  pivot_longer(cols=setdiff(colnames(.),"Gene"),names_to = "Sample.Name")

LI<-left_join(LI_Counts,LI_Meta)%>%
  dplyr::select(-c(Sample.Name))%>%
  pivot_wider(names_from = Parameter,values_from = value)%>%
  rowwise()%>%
  mutate(LI_675_750=100*sd(c(LI675,LI750))/mean(c(LI675,LI750)),
         LI_825_750=100*sd(c(LI825,LI750))/mean(c(LI825,LI750)),
         All=100*sd(c(LI675,LI825,LI750))/mean(c(LI675,LI825,LI750)))%>%
  select(Gene,LI_675_750,LI_825_750,All,Control)%>%
  mutate(Type="Library Input")

Tab<-LI%>%
  select(-c(All,LI_825_750))%>%
  pivot_wider(names_from = Control,values_from = LI_675_750)
```
### Table 6. Batch
```{r}
Batch_Meta<-SampleInfo%>%filter(Category=="Batch")%>%
  dplyr::select(Sample.Name,Control,Parameter)

Batch_Counts<-AllCounts[,Batch_Meta$Sample.Name]%>%
  rownames_to_column(var="Gene")%>%
  pivot_longer(cols=setdiff(colnames(.),"Gene"),names_to = "Sample.Name")

Batch<-left_join(Batch_Counts,Batch_Meta)%>%
  dplyr::select(-c(Sample.Name))%>%
  pivot_wider(names_from = Parameter,values_from = value)%>%
  rowwise()%>%
  mutate(N_25_32=100*sd(c(N25,N32))/mean(c(N25,N32)),
         N_32_39=100*sd(c(N39,N32))/mean(c(N39,N32)),
         All=100*sd(c(N39,N32,N25))/mean(c(N25,N39,N32)))%>%
  select(Gene,N_25_32,N_32_39,All,Control)%>%
  mutate(Type="Batch Size")

Tab<-Batch%>%
  select(-c(All,N_25_32))%>%
  pivot_wider(names_from = Control,values_from = N_32_39)
```

### Table 7. Repeatability
```{r}
Repeat_Meta<-SampleInfo%>%filter(Category=="Repeatability")%>%
  dplyr::select(Sample.Name,Control,Parameter)%>%
  separate(col="Parameter",into = c("Grouping","Trash"),sep = "_")

Repeat_Counts<-AllCounts[,Repeat_Meta$Sample.Name]%>%
  rownames_to_column(var="Gene")%>%
  pivot_longer(cols=setdiff(colnames(.),"Gene"),names_to = "Sample.Name")

Repeat<-left_join(Repeat_Counts,Repeat_Meta)%>%
  dplyr::select(-c(Sample.Name,Trash))%>%
  group_by(Grouping,Gene,Control)%>%
  summarize(Repeat=100*sd(value)/mean(value))%>%
  ungroup()%>%
  group_by(Gene,Control)%>%
  mutate(Mean=mean(Repeat))%>%
  select(Gene,Repeat,Mean,Control,Grouping)%>%
  mutate(Type="Repeatability")%>%
  pivot_wider(names_from = Grouping,values_from = Repeat)

Tab<-Repeat<-left_join(Repeat_Counts,Repeat_Meta)%>%
  dplyr::select(-c(Sample.Name,Trash))%>%
  group_by(Grouping,Gene,Control)%>%
  summarize(Repeat=100*sd(value)/mean(value))%>%
  ungroup()%>%
  mutate(ID=paste(Grouping,Control))%>%
  dplyr::select(-c(Grouping,Control))%>%
  pivot_wider(names_from = ID,values_from = Repeat)
```

### Table 8. Reproducibility
```{r}
Repro_Meta<-SampleInfo%>%filter(Category=="Reproducibility")%>%
  dplyr::select(Sample.Name,Control,Parameter)

Repro_Counts<-AllCounts[,Repro_Meta$Sample.Name]%>%
  rownames_to_column(var="Gene")%>%
  pivot_longer(cols=setdiff(colnames(.),"Gene"),names_to = "Sample.Name")

Repro<-left_join(Repro_Counts,Repro_Meta)%>%
  dplyr::select(-c(Sample.Name))%>%
  pivot_wider(names_from = Parameter,values_from = value)%>%
  rowwise()%>%
  mutate(Reproducibility=100*sd(c(L1R1,L2R1,L1R2,L2R2))/mean(c(L1R1,L2R1,L1R2,L2R2)))%>%
  select(Gene,Control,Reproducibility)%>%
  mutate(Type="Reproducibility")

Tab<-Repro%>%
  pivot_wider(names_from = Control,values_from = Reproducibility)
```

## Combining All tables
```{r}
AllTabs<-bind_rows(FT,Time,DNAse,RI,LI,Batch,Repeat,Repro)%>%
  pivot_longer(cols=setdiff(colnames(.),c("Gene","Control","Type")))%>%
  filter(!is.na(value))

write.csv(AllTabs,"/Users/alinaklineschoder/Desktop/Data_AnalyticalValidation_Organized/CVs.csv")
```

## Plotting the chosen genes summary
```{r}
Filt<-AllTabs%>%
  filter(name%in%c("FT","12mo","DNASE","All","Mean","Reproducibility"))

Filt$Full_Type<-Filt$Type
Filt[Filt$Type%in%c("RNA Stability - Freeze/Thaw","RNA Stability - Time"),"Full_Type"]<-"RNA Stability"
Filt[Filt$Type%in%c("Interference","RNA Input"),"Full_Type"]<-"Library Preparation"
Filt[Filt$Type%in%c("Batch Size","Library Input"),"Full_Type"]<-"Sequencing Characterization"


Filt$Small_Type<-gsub("RNA Stability - ","",Filt$Type)
Filt$Small_Type<-gsub("Time","-80 Storage Time",Filt$Small_Type)
Filt$Small_Type<-gsub("Repeatability","Mean",Filt$Small_Type)

Filt$Full_Type<-factor(Filt$Full_Type,levels=c("RNA Stability","Library Preparation","Sequencing Characterization","Repeatability","Reproducibility"))

Filt$Small_Type<-factor(Filt$Small_Type,levels=rev(c("Freeze/Thaw","-80 Storage Time","Interference","RNA Input","Library Input","Batch Size","Mean","Reproducibility")))


ggplot(Filt, aes(x = Control, fill = value, y = Small_Type)) +
  geom_tile() + theme_minimal()+
  theme(
    legend.position = "right",
    axis.text.x = element_blank(),
    strip.text.y.left = element_text(angle = 0, hjust = 1),
    strip.text.x = element_text(angle = 90, hjust = 1, vjust = 1),
    strip.background = element_blank(),
    axis.ticks.x = element_blank(),panel.grid = element_blank()
  ) +
  scale_fill_gradientn(
    colours = c("navy", "white", "firebrick", "firebrick"),
    values = scales::rescale(c(0, 14.99, 15,20)),  # Extend scale beyond 15
    limits = c(0, max(Genes$value)),  # Adjust to your data range
    oob = scales::squish,  # Keep values >100 at max color
    name = "CV (%)"
  ) +
  facet_grid(Full_Type ~ Gene, scales = "free", space = "free", switch = "both") + theme(axis.title = element_blank())+scale_y_discrete(position="right")

```
